新增理财产品接口

+++ seckill_manager/src/main/java/com/seckill/seckill_manager/entity/FinancialItems.java
@@ -54,7 +54,7 @@ public class FinancialItems implements Serializable {
 
     private Integer shelfLife;
 
-    private Float interestRate;
+    private BigDecimal interestRate;
 
 
     @Override
